Added Location-, Read-in- and Interpolation-Function.
The Location Function 'locate' was taken from the Numerical Recipes in C 
and modfied. ATTENTION there still might be an error of giving back the 
right index.

The 'read-in-data' Function reads data from a file that is passed to it 
("Data/LOOKUPTABLE1.dat" eg.). ATTENTION there still might be an error 
of giving back the right index.

The 'bilinearinterpol' Function is also taken from the Numerical Recipes 
in C - Book. Instructions/Manual will follow!