Bilinear Interpolation and all related things added.
Precalculation(...) calculates the q,z,x_c and f(x_c) or f(x_c_sqrt) 
values and stores them in a file (eg. "Data/LOOKUPTABLE1.dat"). Then 
several pointers/vectors are created which are later passed to the 
read-in-data function to actually "fill" the vectors with data/values. 
Then the function 'locate' gets called to find the correct "square (or 
in other words 4 x-points and 4 f(x) points) to interpolate the desired 
x0 value, which lies in this "square".

This still needs improvement. There is an index wrong in locate or 
read-in-data or something. I manually put in the 'right' values to test 
if the interpolation works. - Worked!