Bugfix and changed dependence of interpol function.
A Factor of 2 was missing in the get_qz - function. After fixing this, 
the interpolation works up to an (accuracy-) epsilon of the 
precalculated function.

bilinearinterpol is now only dependend on q and z (instead of x_c) for 
the sake of brevety. With get_qz we already calculate q and z and can 
directly hand it over to the bilinear-function.