Interpolation of A and B functions added. Also seperate BSE steps.
Finally the Interpolation of the DressingFunctions A and B work.
Had to change double* a_vals and b_vals to std::complex<double>* etc. in
the mother_matrix, theta_matrix, regulaFalsi and bse_root functions.

More commit messages follow in the next couple of days.

TODO:
precalculate the theta_matrix!